HOCS-4277: restructure project
