+++ src/test/java/uk/gov/digital/ho/hocs/client/workflow/WorkflowClientTest.java
@@ -20,9 +20,9 @@ import java.util.UUID;
 
 import static junit.framework.TestCase.assertEquals;
 import static org.mockito.Mockito.when;
-import static uk.gov.digital.ho.hocs.queue.common.ComplaintService.DOCUMENT_TYPE;
-import static uk.gov.digital.ho.hocs.queue.common.ComplaintService.ORIGINAL_FILENAME;
-import static uk.gov.digital.ho.hocs.queue.ukvi.UKVIComplaintService.CASE_TYPE;
+import static uk.gov.digital.ho.hocs.queue.complaints.ComplaintService.DOCUMENT_TYPE;
+import static uk.gov.digital.ho.hocs.queue.complaints.ComplaintService.ORIGINAL_FILENAME;
+import static uk.gov.digital.ho.hocs.queue.complaints.ukvi.UKVIComplaintService.CASE_TYPE;
 
 @RunWith(MockitoJUnitRunner.class)
 public class WorkflowClientTest {
