+++ src/test/java/uk/gov/digital/ho/hocs/queue/complaints/ukvi/UKVIComplaintServiceTest.java
@@ -1,13 +1,11 @@
-package uk.gov.digital.ho.hocs.queue.ukvi;
+package uk.gov.digital.ho.hocs.queue.complaints.ukvi;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.Mock;
 import org.mockito.junit.MockitoJUnitRunner;
 import uk.gov.digital.ho.hocs.application.ClientContext;
-import uk.gov.digital.ho.hocs.queue.common.ComplaintService;
-
-import java.io.IOException;
+import uk.gov.digital.ho.hocs.queue.complaints.ComplaintService;
 
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.eq;
